Centralize option sanitization logic into Service
