+++ tests/php/test-class-amp-options-manager.php
@@ -6,10 +6,7 @@
  */
 
 use AmpProject\AmpWP\Option;
-use AmpProject\AmpWP\PluginRegistry;
-use AmpProject\AmpWP\Services;
 use AmpProject\AmpWP\Tests\AssertContainsCompatibility;
-use AmpProject\AmpWP\Tests\MockPluginEnvironment;
 
 /**
  * Tests for AMP_Options_Manager.
@@ -266,171 +263,6 @@ class Test_AMP_Options_Manager extends WP_UnitTestCase {
 		$entries = AMP_Options_Manager::get_option( Option::ANALYTICS );
 		$this->assertCount( 1, $entries );
 		$this->assertArrayNotHasKey( $id, $entries );
-	}
-
-	/**
-	 * Test validating suppressed plugins.
-	 *
-	 * @covers AMP_Options_Manager::update_option()
-	 * @covers AMP_Options_Manager::validate_suppressed_plugins()
-	 * @covers PluginRegistry::get_plugins()
-	 * @covers AMP_Validated_URL_Post_Type::get_recent_validation_errors_by_source()
-	 */
-	public function test_option_for_suppressed_plugins() {
-		wp_set_current_user( self::factory()->user->create( [ 'role' => 'administrator' ] ) );
-		add_theme_support( AMP_Theme_Support::SLUG );
-		AMP_Validation_Manager::init();
-		AMP_Options_Manager::register_settings(); // Adds validate_options as filter.
-
-		/** @var PluginRegistry $plugin_registry */
-		$plugin_registry = Services::get( 'plugin_registry' );
-
-		// Create bogus plugins.
-		update_option( 'active_plugins', [ MockPluginEnvironment::FOO_PLUGIN_FILE, MockPluginEnvironment::BAR_PLUGIN_FILE, MockPluginEnvironment::AMP_PLUGIN_FILE, MockPluginEnvironment::GUTENBERG_PLUGIN_FILE ] );
-		wp_cache_set( 'plugins', [ '' => MockPluginEnvironment::PLUGINS_DATA ], 'plugins' );
-
-		// Test initial state.
-		$this->assertEquals( [], AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS ) );
-
-		// Test updating empty array.
-		AMP_Options_Manager::update_option( Option::SUPPRESSED_PLUGINS, [] );
-		$this->assertEquals( [], AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS ) );
-
-		$this->assertEmpty( AMP_Validated_URL_Post_Type::get_recent_validation_errors_by_source() );
-
-		$foo_slug = $plugin_registry->get_plugin_slug_from_file( MockPluginEnvironment::FOO_PLUGIN_FILE );
-		$bar_slug = $plugin_registry->get_plugin_slug_from_file( MockPluginEnvironment::BAR_PLUGIN_FILE );
-
-		$errors = [
-			[
-				'code'    => 'removed',
-				'sources' => [
-					[
-						'type' => 'plugin',
-						'name' => $foo_slug,
-					],
-				],
-			],
-			[
-				'code'    => 'kept',
-				'sources' => [
-					[
-						'type' => 'plugin',
-						'name' => $bar_slug,
-					],
-				],
-			],
-			[
-				'code'    => 'removed',
-				'sources' => [
-					[
-						'type' => 'plugin',
-						'name' => $bar_slug,
-					],
-				],
-			],
-		];
-
-		add_filter(
-			'amp_validation_error_default_sanitized',
-			static function( $sanitized, $error ) {
-				if ( 'kept' === $error['code'] ) {
-					$sanitized = false;
-				}
-				return $sanitized;
-			},
-			10,
-			2
-		);
-
-		$r = AMP_Validated_URL_Post_Type::store_validation_errors( $errors, home_url( '/' ) );
-		$this->assertInternalType( 'int', $r );
-
-		$errors_by_source = AMP_Validated_URL_Post_Type::get_recent_validation_errors_by_source();
-		$this->assertEqualSets( [ 'plugin' ], array_keys( $errors_by_source ) );
-		$this->assertEqualSets( [ $foo_slug, $bar_slug ], array_keys( $errors_by_source['plugin'] ) );
-		$foo_plugin_errors = $errors_by_source['plugin'][ $foo_slug ];
-		$bar_plugin_errors = $errors_by_source['plugin'][ $bar_slug ];
-		$this->assertCount( 2, $bar_plugin_errors );
-		$this->assertCount( 1, $foo_plugin_errors );
-
-		$attempted_validate_request_urls = [];
-		add_filter(
-			'pre_http_request',
-			function( $r, $args, $url ) use ( &$attempted_validate_request_urls ) {
-				unset( $r, $args );
-				$attempted_validate_request_urls[] = remove_query_arg( [ 'amp_validate', 'amp_cache_bust' ], $url );
-				return [
-					'body'     => '',
-					'response' => [
-						'code'    => 503,
-						'message' => 'Service Unavailable',
-					],
-				];
-			},
-			10,
-			3
-		);
-
-		// When updating plugins that don't exit or can't be suppressed, do nothing.
-		AMP_Options_Manager::update_option(
-			Option::SUPPRESSED_PLUGINS,
-			[
-				$plugin_registry->get_plugin_slug_from_file( MockPluginEnvironment::GUTENBERG_PLUGIN_FILE ) => '1',
-				$plugin_registry->get_plugin_slug_from_file( MockPluginEnvironment::AMP_PLUGIN_FILE ) => '1',
-				'bogus' => '1',
-			]
-		);
-		remove_all_actions( 'update_option_' . AMP_Options_Manager::OPTION_NAME );
-		$this->assertEquals( [], AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS ) );
-
-		// When updating option but both plugins are not suppressed, then no change is made.
-		AMP_Options_Manager::update_option(
-			Option::SUPPRESSED_PLUGINS,
-			[
-				$foo_slug => '0',
-				$bar_slug => '0',
-			]
-		);
-		remove_all_actions( 'update_option_' . AMP_Options_Manager::OPTION_NAME );
-		$this->assertEquals( [], AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS ) );
-		$this->assertCount( 0, $attempted_validate_request_urls, 'Expected no validation request to have been made since no changes were made (as both plugins are still unsuppressed).' );
-
-		// When updating option and both are now suppressed, then a change is made.
-		AMP_Options_Manager::update_option(
-			Option::SUPPRESSED_PLUGINS,
-			[
-				$foo_slug => '1',
-				$bar_slug => '1',
-			]
-		);
-		remove_all_actions( 'update_option_' . AMP_Options_Manager::OPTION_NAME );
-		$this->assertEquals( array_fill( 0, 1, home_url( '/' ) ), $attempted_validate_request_urls, 'Expected one validation request to have been made since no changes were made (as both plugins are still unsuppressed).' );
-
-		$suppressed_plugins = AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS );
-		$this->assertEqualSets( [ $foo_slug, $bar_slug ], array_keys( $suppressed_plugins ) );
-		foreach ( $suppressed_plugins as $slug => $suppressed_plugin ) {
-			$this->assertArrayHasKey( Option::SUPPRESSED_PLUGINS_LAST_VERSION, $suppressed_plugin );
-			$this->assertArrayHasKey( Option::SUPPRESSED_PLUGINS_TIMESTAMP, $suppressed_plugin );
-			$this->assertArrayHasKey( Option::SUPPRESSED_PLUGINS_USERNAME, $suppressed_plugin );
-			$this->assertEquals( wp_get_current_user()->user_nicename, $suppressed_plugin[ Option::SUPPRESSED_PLUGINS_USERNAME ] );
-			$this->assertArrayHasKey( Option::SUPPRESSED_PLUGINS_ERRORING_URLS, $suppressed_plugin );
-			$this->assertEquals( [ home_url( '/' ) ], $suppressed_plugin[ Option::SUPPRESSED_PLUGINS_ERRORING_URLS ] );
-		}
-		$this->assertEquals( MockPluginEnvironment::PLUGINS_DATA[ MockPluginEnvironment::FOO_PLUGIN_FILE ]['Version'], $suppressed_plugins[ $foo_slug ][ Option::SUPPRESSED_PLUGINS_LAST_VERSION ] );
-		$this->assertEquals( MockPluginEnvironment::PLUGINS_DATA[ MockPluginEnvironment::BAR_PLUGIN_FILE ]['Version'], $suppressed_plugins[ $bar_slug ][ Option::SUPPRESSED_PLUGINS_LAST_VERSION ] );
-
-		// Stop suppressing only one plugin.
-		AMP_Options_Manager::update_option(
-			Option::SUPPRESSED_PLUGINS,
-			[
-				$foo_slug => '1',
-				$bar_slug => '0',
-			]
-		);
-		remove_all_actions( 'update_option_' . AMP_Options_Manager::OPTION_NAME );
-		$this->assertEquals( array_fill( 0, 2, home_url( '/' ) ), $attempted_validate_request_urls, 'Expected one validation request to have been made since no changes were made (as both plugins are still unsuppressed).' );
-		$this->assertEqualSets( [ $foo_slug ], array_keys( AMP_Options_Manager::get_option( Option::SUPPRESSED_PLUGINS ) ) );
 	}
 
 	/**
